Final_Assignment/Final-Assignment_code/src/main/java/nuig/oop/model/Films.java
package nuig.oop.model;

import com.fasterxml.jackson.annotation.JsonIdentityInfo;
import com.fasterxml.jackson.annotation.ObjectIdGenerators;

@JsonIdentityInfo(generator = ObjectIdGenerators.PropertyGenerator.class, property = "fid", scope = Films.class)
public class Films extends Media {
	private int fid;
	private String director;

	private Films() {
	}

	public static Films getInstance() {
		return new Films();
	}

	public int getFid() {
		return fid;
	}

	public void setFid(int fid) {
		this.fid = fid;
	}

	public String getDirector() {
		return director;
	}

	public void setDirector(String director) {
		this.director = director;
	}

}
